update logdata format
