+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-common/src/main/java/org/apache/skywalking/apm/toolkit/common/log/GRPCLogConfig.java
@@ -41,11 +41,6 @@ public class GRPCLogConfig {
             public static int MAX_MESSAGE_SIZE = 10 * 1024 * 1024;
 
             /**
-             * The log body type to send to the server, includes text(2),json(3),yaml(4).Default is text.
-             */
-            public static int LOG_BODY_TYPE = 2;
-
-            /**
              * How long grpc client will timeout in sending data to upstream.
              */
             public static int UPSTREAM_TIMEOUT = 30;
