+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-common/src/main/resources/META-INF/services/org.apache.skywalking.apm.agent.core.boot.BootService
@@ -19,4 +19,4 @@
 #
 #
 
-org.apache.skywalking.apm.toolkit.common.log.LogReportServiceClient
+org.apache.skywalking.apm.toolkit.common.log.GRPCLogReportServiceClient
