+++ docs/en/setup/service-agent/java-agent/Application-toolkit-log4j-1.x.md
@@ -18,3 +18,32 @@ log4j.appender.CONSOLE.layout.ConversionPattern=%d [%T] %-5p %c{1}:%L - %m%n
 ```
 
 * When you use `-javaagent` to active the sky-walking tracer, log4j will output **traceId**, if it existed. If the tracer is inactive, the output will be `TID: N/A`.
+
+# grpc log log4j plugin
+
+The grpc log plugin support report log to grpc Server. 
+
+* Dependency the toolkit, such as using maven or gradle
+
+```xml
+   <dependency>
+      <groupId>org.apache.skywalking</groupId>
+      <artifactId>apm-toolkit-log4j-1.x</artifactId>
+      <version>{project.release.version}</version>
+   </dependency>
+```
+
+* add `GRPCLogClientAppender` in logback.xml
+
+```xml
+    <appender name="grpc-log" class="org.apache.skywalking.apm.toolkit.log.log4j.v1.x.log.GRPCLogClientAppender">
+    </appender>
+```
+
+*  modify config of the plugin 
+```properties
+plugin.grpclog.service_host=${SW_GRPC_LOG_SERVER_HOST:127.0.0.1}
+plugin.grpclog.service_port=${SW_GRPC_LOG_SERVER_PORT:8000}
+plugin.grpclog.max_message_size=${SW_GRPC_LOG_MAX_MESSAGE_SIZE:10485760}
+plugin.grpclog.upstream_timeout=${SW_GRPC_LOG_GRPC_UPSTREAM_TIMEOUT:30}
+```
